requirements.txt
pymobiledevice3
python-socketio
eventlet
